Feature/etl rest tests

Created backend tests for ETL REST using TestNG and Jersey Test Framework. Changed Mapping and CSV REST code to return null values and use 204 status codes instead of 400 status codes. Updated exceptions being thrown by the CSVConverter.

See merge request !149

+++ org.matonto.etl.rest/src/main/java/org/matonto/etl/rest/impl/CSVRestImpl.java
@@ -11,6 +11,7 @@ import org.glassfish.jersey.media.multipart.FormDataContentDisposition;
 import org.matonto.etl.api.csv.CSVConverter;
 import org.matonto.etl.api.csv.MappingManager;
 import org.matonto.etl.rest.CSVRest;
+import org.matonto.exception.MatOntoException;
 import org.matonto.rdf.api.Resource;
 import org.matonto.rdf.core.utils.Values;
 import org.matonto.rest.util.ErrorUtils;
@@ -28,7 +29,9 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
-import java.util.*;
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
 import javax.ws.rs.core.Response;
 
 @Component(immediate = true)
@@ -61,13 +64,20 @@ public class CSVRestImpl implements CSVRest {
 
     @Override
     public Response upload(InputStream fileInputStream, FormDataContentDisposition fileDetail, String fileName) {
-        File delimitedFile = getUploadedFile(fileName);
         String newExtension = FilenameUtils.getExtension(fileDetail.getFileName());
-        if (!newExtension.equals(FilenameUtils.getExtension(delimitedFile.getName()))) {
-            delimitedFile.delete();
-            delimitedFile = new File("data/tmp/" + fileName + "." + newExtension);
+        Optional<File> optDelimitedFile = getUploadedFile(fileName);
+        Path filePath;
+        if (optDelimitedFile.isPresent()) {
+            File delimitedFile = optDelimitedFile.get();
+            if (!newExtension.equals(FilenameUtils.getExtension(delimitedFile.getName()))) {
+                delimitedFile.delete();
+                delimitedFile = new File("data/tmp/" + fileName + "." + newExtension);
+            }
+            filePath = delimitedFile.toPath();
+        } else {
+            filePath = Paths.get("data/tmp/" + fileName + "." + newExtension);
         }
-        uploadFile(fileInputStream, delimitedFile.toPath());
+        uploadFile(fileInputStream, filePath);
         return Response.status(200).entity(fileName).build();
     }
 
@@ -79,75 +89,85 @@ public class CSVRestImpl implements CSVRest {
                     Response.Status.BAD_REQUEST);
         }
 
-        File delimitedFile = getUploadedFile(fileName);
-        String extension = FilenameUtils.getExtension(delimitedFile.getName());
-        char separatorChar = separator.charAt(0);
+        Optional<File> optDelimitedFile = getUploadedFile(fileName);
+        if (optDelimitedFile.isPresent()) {
+            File delimitedFile = optDelimitedFile.get();
+            String extension = FilenameUtils.getExtension(delimitedFile.getName());
+            char separatorChar = separator.charAt(0);
 
-        // Get InputStream for data to convert
-        InputStream dataToConvert;
-        if (isPreview) {
-            dataToConvert = (extension.equals("xls") || extension.equals("xlsx"))
-                    ? createExcelPreviewStream(delimitedFile, containsHeaders)
-                    : createCSVPreviewStream(delimitedFile, containsHeaders);
-        } else {
-            try {
-                dataToConvert = new FileInputStream(delimitedFile);
-            } catch (FileNotFoundException e) {
-                throw ErrorUtils.sendError(e, "Error locating delimited file", Response.Status.BAD_REQUEST);
-            }
-        }
-
-        // Convert InputStream to RDF based on Mapping
-        Model model;
-        try {
-            Model mappingModel;
-            if (mappingLocalName != null) {
-                Resource mappingIRI = mappingManager.createMappingIRI(mappingLocalName);
-                Optional<org.matonto.rdf.api.Model> mappingOptional = mappingManager.retrieveMapping(mappingIRI);
-                if (mappingOptional.isPresent()) {
-                    mappingModel = Values.sesameModel(mappingOptional.get());
-                } else {
-                    throw ErrorUtils.sendError("Mapping " + mappingIRI + " does not exist",
-                            Response.Status.BAD_REQUEST);
-                }
+            // Get InputStream for data to convert
+            InputStream dataToConvert;
+            if (isPreview) {
+                dataToConvert = (extension.equals("xls") || extension.equals("xlsx"))
+                        ? createExcelPreviewStream(delimitedFile, containsHeaders)
+                        : createCSVPreviewStream(delimitedFile, containsHeaders);
             } else {
-                InputStream in = new ByteArrayInputStream(mappingRdf.getBytes(StandardCharsets.UTF_8));
-                mappingModel = Rio.parse(in, "", RDFFormat.JSONLD);
+                try {
+                    dataToConvert = new FileInputStream(delimitedFile);
+                } catch (FileNotFoundException e) {
+                    throw ErrorUtils.sendError(e, "Error locating delimited file", Response.Status.BAD_REQUEST);
+                }
             }
-            model = Values.sesameModel(csvConverter.convert(dataToConvert,
-                    Values.matontoModel(mappingModel), containsHeaders, extension, separatorChar));
-        } catch (IOException | InvalidFormatException e) {
-            throw ErrorUtils.sendError(e, "Error converting delimited file", Response.Status.BAD_REQUEST);
-        }
 
-        // Write data back to Response
-        logger.info("File mapped: " + delimitedFile.getPath());
-        StringWriter sw = new StringWriter();
-        RDFHandler rdfWriter = new BufferedGroupingRDFHandler(Rio.createWriter(getRDFFormat(format), sw));
-        Rio.write(model, rdfWriter);
-        return Response.status(200).entity(sw.toString()).build();
+            // Convert InputStream to RDF based on Mapping
+            Model model;
+            try {
+                Model mappingModel;
+                if (mappingLocalName != null) {
+                    Resource mappingIRI = mappingManager.createMappingIRI(mappingLocalName);
+                    Optional<org.matonto.rdf.api.Model> mappingOptional = mappingManager.retrieveMapping(mappingIRI);
+                    if (mappingOptional.isPresent()) {
+                        mappingModel = Values.sesameModel(mappingOptional.get());
+                    } else {
+                        throw ErrorUtils.sendError("Mapping " + mappingIRI + " does not exist",
+                                Response.Status.BAD_REQUEST);
+                    }
+                } else {
+                    InputStream in = new ByteArrayInputStream(mappingRdf.getBytes(StandardCharsets.UTF_8));
+                    mappingModel = Rio.parse(in, "", RDFFormat.JSONLD);
+                }
+                model = Values.sesameModel(csvConverter.convert(dataToConvert,
+                        Values.matontoModel(mappingModel), containsHeaders, extension, separatorChar));
+            } catch (IOException | MatOntoException e) {
+                throw ErrorUtils.sendError(e, "Error converting delimited file", Response.Status.BAD_REQUEST);
+            }
+
+            // Write data back to Response
+            logger.info("File mapped: " + delimitedFile.getPath());
+            StringWriter sw = new StringWriter();
+            RDFHandler rdfWriter = new BufferedGroupingRDFHandler(Rio.createWriter(getRDFFormat(format), sw));
+            Rio.write(model, rdfWriter);
+            return Response.status(200).entity(sw.toString()).build();
+        } else {
+            throw ErrorUtils.sendError("Document not found", Response.Status.BAD_REQUEST);
+        }
     }
 
     @Override
     public Response getRows(String fileName, int rowEnd, String separator) {
-        File file = getUploadedFile(fileName);
-        String extension = FilenameUtils.getExtension(file.getName());
-        int numRows = (rowEnd <= 0) ? 10 : rowEnd;
+        Optional<File> optFile = getUploadedFile(fileName);
+        if (optFile.isPresent()) {
+            File file = optFile.get();
+            String extension = FilenameUtils.getExtension(file.getName());
+            int numRows = (rowEnd <= 0) ? 10 : rowEnd;
 
-        logger.info("Getting " + numRows + " rows from " + file.getName());
-        String json;
-        try {
-            if (extension.equals("xls") || extension.equals("xlsx")) {
-                json = convertExcelRows(file, numRows);
-            } else {
-                char separatorChar = separator.charAt(0);
-                json = convertCSVRows(file, numRows, separatorChar);
+            logger.info("Getting " + numRows + " rows from " + file.getName());
+            String json;
+            try {
+                if (extension.equals("xls") || extension.equals("xlsx")) {
+                    json = convertExcelRows(file, numRows);
+                } else {
+                    char separatorChar = separator.charAt(0);
+                    json = convertCSVRows(file, numRows, separatorChar);
+                }
+            } catch (Exception e) {
+                throw ErrorUtils.sendError("Error loading document", Response.Status.BAD_REQUEST);
             }
-        } catch (Exception e) {
-            throw ErrorUtils.sendError("Error loading document", Response.Status.BAD_REQUEST);
-        }
 
-        return Response.status(200).entity(json).build();
+            return Response.status(200).entity(json).build();
+        } else {
+            throw ErrorUtils.sendError("Document not found", Response.Status.BAD_REQUEST);
+        }
     }
 
     /**
@@ -156,19 +176,19 @@ public class CSVRestImpl implements CSVRest {
      * @param fileName the name of the uploaded delimited file
      * @return the uploaded file if it was found
      */
-    private File getUploadedFile(String fileName) {
+    private Optional<File> getUploadedFile(String fileName) {
         File directory = new File("data/tmp");
         File[] files = directory.listFiles((dir, name) -> {
             return name.startsWith(fileName + ".");
         });
         if (files.length == 0) {
-            throw ErrorUtils.sendError("Delimited file does not exist", Response.Status.BAD_REQUEST);
+            return Optional.empty();
         }
         if (files.length > 1) {
             throw ErrorUtils.sendError("Multiple files exist with same name", Response.Status.BAD_REQUEST);
         }
 
-        return files[0];
+        return Optional.of(files[0]);
     }
 
     /**
@@ -179,7 +199,7 @@ public class CSVRestImpl implements CSVRest {
      */
     private RDFFormat getRDFFormat(String format) {
         RDFFormat rdfformat;
-        switch (format) {
+        switch (format.toLowerCase()) {
             case "turtle":
                 rdfformat = RDFFormat.TURTLE;
                 break;
